refactor: client view survey page
Fixes: #1111
