+++ src/app/clients/clients-routing.module.ts
@@ -20,7 +20,6 @@ import { DocumentsTabComponent } from './clients-view/documents-tab/documents-ta
 import { DatatableTabComponent } from './clients-view/datatable-tab/datatable-tab.component';
 import { AddressTabComponent } from './clients-view/address-tab/address-tab.component';
 import { ClientActionsComponent } from './clients-view/client-actions/client-actions.component';
-import { TakeSurveyComponent } from './clients-view/client-actions/view-survey/take-survey/take-survey.component';
 
 /** Custom Resolvers */
 import { ClientViewResolver } from './common-resolvers/client-view.resolver';
@@ -42,10 +41,7 @@ import { ClientAddressTemplateResolver } from './common-resolvers/client-address
 import { ChargesOverviewComponent } from './clients-view/charges-overview/charges-overview.component';
 import { ClientChargeOverviewResolver } from './clients-view/charges-overview/charge-overview.resolver';
 import { ClientActionsResolver } from './common-resolvers/client-actions.resolver';
-import { TakeSurveyResolver } from './common-resolvers/take-survey.resolver';
-import { ViewChargeComponent } from './clients-view/view-charge/view-charge.component';
 import { ClientChargeViewResolver } from './common-resolvers/client-charge-view.resolver';
-import { ClientPayChargesComponent } from './clients-view/client-pay-charges/client-pay-charges.component';
 import { ClientTransactionPayResolver } from './common-resolvers/client-transaction-pay.resolver';
 
 const routes: Routes = [
@@ -179,23 +175,10 @@ const routes: Routes = [
           {
             path: 'actions/:name',
             data: { title: extract('Client Actions'), routeParamBreadcrumb: 'name' },
-            children: [
-              {
-                path: '',
-                component: ClientActionsComponent,
-                resolve: {
-                  clientActionData: ClientActionsResolver
-                }
-              },
-              {
-                path: 'Take-Survey',
-                component: TakeSurveyComponent,
-                data: { title: extract('Take Survey'), breadcrumb: 'Take Survey', routeParamBreadcrumb: 'Survey' },
-                resolve: {
-                  allSurveysData: TakeSurveyResolver
-                }
-              },
-            ]
+            component: ClientActionsComponent,
+            resolve: {
+              clientActionData: ClientActionsResolver
+            }
           },
           {
             path: 'loans',
@@ -247,7 +230,6 @@ const routes: Routes = [
     ClientAddressTemplateResolver,
     ClientChargeOverviewResolver,
     ClientActionsResolver,
-    TakeSurveyResolver,
     ClientChargeViewResolver,
     ClientTransactionPayResolver
   ]
