+++ src/app/clients/clients-view/client-actions/client-actions.component.html
@@ -12,3 +12,4 @@
 <mifosx-reactivate-client *ngIf="actions['Reactivate']"></mifosx-reactivate-client>
 <mifosx-undo-client-rejection *ngIf="actions['Undo Rejection']"></mifosx-undo-client-rejection>
 <mifosx-add-client-charge *ngIf="actions['Add Charge']"></mifosx-add-client-charge>
+<mifosx-take-survey *ngIf="actions['Take Survey']"></mifosx-take-survey>
