+++ src/app/clients/clients-view/client-actions/take-survey/take-survey.component.html
@@ -1,7 +1,5 @@
 <div class="container mat-typography" style="padding-bottom: 20px;">
 
-  <h2>Surveys</h2>
-
   <div class="mat-elevation-z2 surveyCont">
 
     <div class="surveySelect" fxLayout="row">
