+++ src/app/clients/clients-view/client-actions/take-survey/take-survey.component.scss
